Breaking out AbstractGuardrailPlugin
